requirements.txt
soundfile
sounddevice
pynput
pillow
numpy
torch
mss
easyocr
lingua-language-detector
argostranslate
omnivoice
PySide6
jsonmerge
jsonref
pyjson5
addict
ollama
faster-whisper
requests
werkzeug
flask
scipy